feat-walk/refactor: 백드롭 컴포넌트 분리
